+++ src/QueryFactory.php
@@ -4,15 +4,11 @@ declare(strict_types=1);
 namespace Latitude\QueryBuilder;
 
 use Latitude\QueryBuilder\Ruler\Context;
-use Latitude\QueryBuilder\Ruler\Visitor;
 
 class QueryFactory
 {
     /** @var EngineInterface */
     protected $engine;
-
-    /** @var Visitor */
-    protected $visitor;
 
     public function __construct(
         EngineInterface $engine = null
@@ -66,7 +62,6 @@ class QueryFactory
     public function delete($table): Query\DeleteQuery
     {
         return $this->engine->makeDelete()->from($table);
-
     }
 
     /**
@@ -81,24 +76,5 @@ class QueryFactory
             $query = $query->set($map);
         }
         return $query;
-    }
-
-    /**
-     * Create a new criteria from an expression
-     *
-     * @link https://hoa-project.net/En/Literature/Hack/Ruler.html
-     */
-    public function criteria(string $expression): CriteriaInterface
-    {
-        return $this->visitor()->visit(\Hoa\Ruler\Ruler::interpret($expression));
-    }
-
-    protected function visitor(): Visitor
-    {
-        if ($this->visitor === null) {
-            $this->visitor = new Visitor();
-        }
-
-        return $this->visitor;
     }
 }